Include IPAddresses in the script and removed from code. Also updated README

+++ start-experiment.sh
@@ -2,6 +2,10 @@
 
 # Output of Experiment logs will be written directly to the project Directory experiment-logs
 runExperiments() {
+
+  # Store IP addresses in a list
+  allIPAddr=($4 $5 $6 $7 $8)
+
   # Remove experiment-logs directory and create new one
   echo "Create new experiment-logs directory in project structure"
   sshpass -p $1 ssh cs4224j@xcnc20.comp.nus.edu.sg "source .bash_profile; rm -rf temp/cockroachdb/src/main/java/experiment-logs; mkdir temp/cockroachdb/src/main/java/experiment-logs"
@@ -14,10 +18,11 @@ runExperiments() {
   for ((i=0; i<$2; i++)); do
     logIndex=$(($i + 1)) # Indexed 1
     # $3 contains the number of servers to run on
-    nodeid=$(($i % $3)) # Used to pass inside Main Function so we know which node's IP address to connect to
+    nodeid=$(($i % $3))
+    serverIP=${allIPAddr[nodeid]} # Used to pass inside Main Function so we know which node's IP address to connect to
     serverID=$((20 + nodeid))
   	server="xcnc${serverID}"
-  	echo "Assign client $i on $server" # client Indexed 0
+  	echo "Assign client $(($i + 1)) on $server with IP Address being $serverIP" # client Indexed 1
 
   	input_file="src/main/java/DataSource/xact-files/${logIndex}.txt"
   	stdout_file="src/main/java/experiment-logs/${logIndex}.out.log"
@@ -28,13 +33,13 @@ runExperiments() {
   	sshpass -p $1 ssh cs4224j@${server}.comp.nus.edu.sg "source .bash_profile; touch temp/cockroachdb/src/main/java/experiment-logs/${logIndex}.txt; touch temp/cockroachdb/src/main/java/experiment-logs/${logIndex}.out.log; touch temp/cockroachdb/src/main/java/experiment-logs/${logIndex}.err.log"
 
   	echo "Start running transactions via Main function"
-    sshpass -p $1 ssh cs4224j@${server}.comp.nus.edu.sg "source .bash_profile; cd temp/cockroachdb && java -Xms4g -Xmx4g -cp target/*:target/dependency/*:. Main $nodeid < ${input_file} > ${stdout_file} 2> ${stderr_file} &"> /dev/null 2>&1 &
+    sshpass -p $1 ssh cs4224j@${server}.comp.nus.edu.sg "source .bash_profile; cd temp/cockroachdb && java -Xms4g -Xmx4g -cp target/*:target/dependency/*:. Main $serverIP < ${input_file} > ${stdout_file} 2> ${stderr_file} &"> /dev/null 2>&1 &
   	echo "Finish running ${i+1} transaction file on $server"
   done
 }
 
 loadDataToDatabaseFromExternFolder() {
-  sshpass -p $1 ssh cs4224j@xcnc20.comp.nus.edu.sg "source .bash_profile; cd temp/cockroachdb && mvn clean dependency:copy-dependencies package; java -Xms4g -Xmx4g -cp target/*:target/dependency/*:. InitialiseData"
+  sshpass -p $1 ssh cs4224j@xcnc20.comp.nus.edu.sg "source .bash_profile; cd temp/cockroachdb && mvn clean dependency:copy-dependencies package; java -Xms4g -Xmx4g -cp target/*:target/dependency/*:. InitialiseData $2"
   echo "Built project on xcnc20"
 }
 
@@ -42,7 +47,6 @@ loadDataToDatabaseFromExternFolder() {
 createExternFolderToStoreCSV() {
   echo "Create new /extern directories for each node"
   echo $1
-  # sshpass -p "$1" ssh cs4224j@xcnc20.comp.nus.edu.sg "source .bash_profile; mkdir temp/node1/extern/; cp -r temp/cockroachdb/src/main/java/DataSource/* ~/temp/node1/extern/;"
   sshpass -p $1 ssh cs4224j@xcnc20.comp.nus.edu.sg "source .bash_profile; rm -rf temp/node1/extern/; mkdir temp/node1/extern/; cp -r temp/cockroachdb/src/main/java/DataSource/* ~/temp/node1/extern/;"
   echo "node 1"
   sshpass -p $1 ssh cs4224j@xcnc21.comp.nus.edu.sg "source .bash_profile; rm -rf temp/node2/extern/; mkdir temp/node2/extern/; cp -r temp/cockroachdb/src/main/java/DataSource/* ~/temp/node2/extern/;"
@@ -51,23 +55,30 @@ createExternFolderToStoreCSV() {
   echo "node 3"
   sshpass -p $1 ssh cs4224j@xcnc23.comp.nus.edu.sg "source .bash_profile; rm -rf temp/node4/extern/; mkdir temp/node4/extern/; cp -r temp/cockroachdb/src/main/java/DataSource/* ~/temp/node4/extern/;"
   echo "node 4"
+
+  if [ $2 == 5 ]
+  then
   sshpass -p $1 ssh cs4224j@xcnc24.comp.nus.edu.sg "source .bash_profile; rm -rf temp/node5/extern/; mkdir temp/node5/extern/; cp -r temp/cockroachdb/src/main/java/DataSource/* ~/temp/node5/extern/;"
   echo "node 5"
+  fi
   echo "Created /extern directory for all nodes and transfer CSV files inside each directory"
 }
 
 
 
 # ============================START=============================
-## $1: Password, $2 Number of clients: 20/40, $3 number of servers: 4/5
+## $1: Password, $2 Number of clients: 20/40, $3 number of servers: 4/5, $4 $5 $6 $7 $8 servers IP Addresses
 echo "Starting to load data-files into all nodes' /extern directory"
-createExternFolderToStoreCSV $1
+createExternFolderToStoreCSV $1 $3
 
+# We load data into the first server address given
 echo "Load data-files into cockroachdb database.....call InitialiseData File in the Project"
-loadDataToDatabaseFromExternFolder $1
+loadDataToDatabaseFromExternFolder $1 $4
 
-echo "Start Experiment. Please key in parameters with the first being the number of clients followed by number of servers"
+echo "Start Experiment. Please key in parameters with the first being the number of clients followed by number of servers and then IP addresses of the servers to use"
+echo "Example Usage: password 20 5 ipOne ipTwo ipThree ipFour ipFive"
 echo "starting to run with $2 number of clients"
 echo "Starting to run with $3 number of servers"
-runExperiments $1 $2 $3
+echo "Server IP Addresses are $4 $5 $6 $7 $8"
+runExperiments $1 $2 $3 $4 $5 $6 $7 $8
 echo "Complete experiment"
